% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freeVenn.R
\name{freeVenn}
\alias{freeVenn}
\title{Make Venn diagrams for two sets, three sets, and four sets.}
\usage{
freeVenn(vlist, ...)
}
\arguments{
\item{vlist}{a list with proper names}

\item{...}{all other parameters used for customizing plot}
}
\value{
a figure as ggplot2 object
}
\description{
Make Venn diagrams for two sets, three sets, and four sets.
}
\examples{
vlist <- list(groupA=sample(1:100, 80), groupB=sample(50:150, 70), groupC=sample(1:100, 60))
freeVenn(vlist)
## rough Weighted
freeVenn(vlist, weighted=TRUE)
## optimize Weighted
freeVenn(vlist, weighted=TRUE, Optimize=TRUE)

}
